+++ src/bstone/src/id_ca.cpp
@@ -40,6 +40,7 @@ loaded into the data segment
 #include "bstone_bmp_image_common.h"
 #include "bstone_endian.h"
 #include "bstone_exception.h"
+#include "bstone_fs.h"
 #include "bstone_fs_utils.h"
 #include "bstone_globals.h"
 #include "bstone_image_extractor.h"
@@ -1748,6 +1749,7 @@ void ca_extract_music(
 	auto audio_extractor = bstone::make_audio_extractor(*audio_content_mgr);
 
 	const auto normalized_dst_dir = bstone::fs_utils::normalize_path(destination_dir);
+	bstone::fs::create_directories(normalized_dst_dir.c_str());
 	audio_extractor->extract_music(normalized_dst_dir);
 
 	bstone::globals::logger->log_information(">>> ================");
@@ -1765,6 +1767,7 @@ void ca_extract_sfx(
 	auto audio_extractor = bstone::make_audio_extractor(*audio_content_mgr);
 
 	const auto normalized_dst_dir = bstone::fs_utils::normalize_path(destination_dir);
+	bstone::fs::create_directories(normalized_dst_dir.c_str());
 	audio_extractor->extract_sfx(normalized_dst_dir);
 
 	bstone::globals::logger->log_information(">>> ================");
@@ -1779,6 +1782,7 @@ void ca_extract_texts(
 	bstone::globals::logger->log_information(("Destination dir: \"" + destination_dir + "\"").c_str());
 
 	auto text_extractor = bstone::TextExtractor{};
+	bstone::fs::create_directories(destination_dir.c_str());
 	text_extractor.extract_text(bstone::fs_utils::normalize_path(destination_dir));
 
 	bstone::globals::logger->log_information(">>> ================");